update java
Update to lib version with shex schema parser fix.  Now also works with the shapemap and the subclass-based selector pattern to find the focus nodes.  Input files must be enriched with the superclass relations for all the classes we care about (e.g. those that will hit the roots in the shape map).  Unit Test is working.