not going to relax BP part of BP for now
Adding the skos relations to make it possible to capture hierarchical node-node relations outside the context of the OWL inference system and the current Noctua infrastructure.  These are useful for e.g. the reactome pathway hierarchy - allowing us access to it from SPARQL, but stopping it from confusing models or adding unwanted inferences.